movie/serializers.py
from rest_framework import serializers
# from rest_framework.authtoken.admin import User
from django.contrib.auth import get_user_model

from .models import Movie, Author, Genre, CustomUser

User = get_user_model()



class AuthorSerializerFromForeign(serializers.ModelSerializer):
    class Meta:
        model = Author
        fields = ['name', 'birthDate']

class MovieSerializer(serializers.ModelSerializer):
    class Meta:
        model = Movie
        fields = ['title', 'publishedDate', 'author', 'genre']

class AuthorSerializer(serializers.ModelSerializer):
    class Meta:
        model = Author
        fields = '__all__'

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = '__all__'

class RegisterSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = ['username', 'email', 'password', 'enabled']
        extra_kwargs = {'password': {'write_only': True}}

    def create(self, validated_data):
        user = CustomUser.objects.create_user(
            username=validated_data['username'],
            email=validated_data['email']
        )
        user.set_password(validated_data['password'])
        user.save()
        return user

class GenreSerializer(serializers.ModelSerializer):
    class Meta:
        model = Genre
        fields = '__all__'